+++ llm_guard/input_scanners/toxicity.py
@@ -1,5 +1,7 @@
+from enum import Enum
+
 from llm_guard.transformers_helpers import pipeline
-from llm_guard.util import calculate_risk_score, logger
+from llm_guard.util import calculate_risk_score, logger, split_text_by_sentences
 
 from .base import Scanner
 
@@ -14,7 +16,13 @@ _toxic_labels = [
     "threat",
     "insult",
     "identity_attack",
+    "sexual_explicit",
 ]
+
+
+class MatchType(Enum):
+    SENTENCE = "sentence"
+    FULL = "full"
 
 
 class Toxicity(Scanner):
@@ -25,12 +33,15 @@ class Toxicity(Scanner):
     considered toxic.
     """
 
-    def __init__(self, threshold: float = 0.7, use_onnx: bool = False):
+    def __init__(
+        self, threshold: float = 0.5, match_type: MatchType = MatchType.FULL, use_onnx: bool = False
+    ):
         """
         Initializes Toxicity with a threshold for toxicity.
 
         Parameters:
-           threshold (float): Threshold for toxicity. Default is 0.7.
+           threshold (float): Threshold for toxicity. Default is 0.5.
+           match_type (MatchType): Whether to match the full text or individual sentences. Default is MatchType.FULL.
            use_onnx (bool): Whether to use ONNX for inference. Default is False.
 
         Raises:
@@ -38,12 +49,16 @@ class Toxicity(Scanner):
         """
 
         self._threshold = threshold
+        self._match_type = match_type
+
         self._pipeline = pipeline(
             task="text-classification",
             model=_model_path[0],
             onnx_model=_model_path[1],
             top_k=None,
             use_onnx=use_onnx,
+            padding="max_length",
+            function_to_apply="sigmoid",
             truncation=True,
         )
 
@@ -51,24 +66,29 @@ class Toxicity(Scanner):
         if prompt.strip() == "":
             return prompt, True, 0.0
 
-        results = self._pipeline(prompt)
+        inputs = [prompt]
+        if self._match_type == MatchType.SENTENCE:
+            inputs = split_text_by_sentences(prompt)
 
         highest_toxicity_score = 0.0
         toxicity_above_threshold = []
-        for result in results[0]:
-            if result["label"] not in _toxic_labels:
-                continue
+        results_all = self._pipeline(inputs)
+        for results_chunk in results_all:
+            for result in results_chunk:
+                if result["label"] not in _toxic_labels:
+                    continue
 
-            if result["score"] > self._threshold:
-                toxicity_above_threshold.append(result)
-            if result["score"] > highest_toxicity_score:
-                highest_toxicity_score = result["score"]
+                if result["score"] > self._threshold:
+                    toxicity_above_threshold.append(result)
+
+                if result["score"] > highest_toxicity_score:
+                    highest_toxicity_score = result["score"]
 
         if len(toxicity_above_threshold) > 0:
             logger.warning(f"Detected toxicity in the text: {toxicity_above_threshold}")
 
             return prompt, False, calculate_risk_score(highest_toxicity_score, self._threshold)
 
-        logger.debug(f"Not toxicity found in the text. Results: {results}")
+        logger.debug(f"Not toxicity found in the text. Results: {results_all}")
 
         return prompt, True, 0.0