+++ docs/changelog.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **[Breaking]** `Regex` scanners have a new signature. It accepts `patterns`, `is_blocked` and `match_type`.
 - **[Breaking]** `BanSubstrings`: `match_type` parameter became `Enum` instead of `str`.
 - **[Breaking]** `Code` scanners have a new signature. It accepts `languages` and `is_blocked` instead of 2 separate lists.
+- `Toxicity` scanner supports sentence match for better accuracy (will become slower).
 
 ### Removed
 -